+++ src/Controllers/PlacesController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Sfneal\GooglePlaces;
+namespace Sfneal\GooglePlaces\Controllers;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Sfneal\Controllers\AbstractController;
+use Sfneal\Controllers\Controller;
+use Sfneal\GooglePlaces\Services\Autocomplete;
 
-class PlacesController extends AbstractController
+class PlacesController extends Controller
 {
     /**
      * Execute a Google Place API city search.