+++ src/Services/Autocomplete.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Sfneal\GooglePlaces;
+namespace Sfneal\GooglePlaces\Services;
 
-use Sfneal\Actions\AbstractService;
 use Sfneal\GooglePlaces\Actions\AutocompleteCityAction;
 use Sfneal\GooglePlaces\Actions\AutocompleteZipAction;
 
-class Autocomplete extends AbstractService
+class Autocomplete
 {
     /**
      * Autocomplete a 'city' name query.