+++ CHANGELOG.md
@@ -57,3 +57,9 @@ All notable changes to `google-places` will be documented in this file
 ## 1.0.1 - 2021-03-30
 - fix sfneal/actions & sfneal/controllers version syntax
 - fix Travis CI config to enable code coverage uploads
+
+
+## 1.1.0 - 2021-03-31
+- bump sfneal/actions & sfneal/controllers min versions to 2.0
+- refactor `Autocomplete` & `Extract` into Services namespace
+- refactor `PlacesController` into Controllers namespace